refactor: 타입 파일 네이밍 컨벤션 통일 (*.types.ts → *.type.ts)"
