+++ src/lib/api/mover/requests/getMover.ts
@@ -1,8 +1,7 @@
 //lib/api/mover/getMover.ts
 import { tokenFetch, defaultFetch } from "@/lib/utils/fetch-client";
-import { Mover } from "@/lib/types/auth.types";
-import { GetMoversParams, GetMoversResponse } from "@/lib/types/mover.types";
 import { delay } from "../../../../../delay";
+import { GetMoversParams, GetMoversResponse, Mover } from "@/lib/types";
 
 export const getMovers = async (
    params: GetMoversParams = {},
