+++ src/lib/hooks/useClientProfilePostForm.ts
@@ -7,8 +7,7 @@ import { useForm } from "react-hook-form";
 import { ClientProfilePostValue, useClientProfileSchemas } from "../schemas";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { uploadToCloudinary } from "../api/auth/requests/updateProfileImage";
-import { AuthFetchError } from "../types";
-import { ServiceType } from "../types/client.types";
+import { AuthFetchError, ServiceType } from "../types";
 import clientProfile from "../api/auth/requests/updateClientProfile";
 import { tokenSettings } from "../utils";
 import { useToast } from "@/context/ToastContext";
