+++ src/lib/hooks/useClientProfileUpdateForm.ts
@@ -5,10 +5,9 @@ import { useEffect, useMemo, useState } from "react";
 import { useAuth } from "@/context/AuthContext";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { AuthFetchError, Client, User } from "../types";
+import { AuthFetchError, Client, ServiceType, User } from "../types";
 import { ClientProfileUpdateValue, useClientProfileSchemas } from "../schemas";
 import clientProfile from "../api/auth/requests/updateClientProfile";
-import { ServiceType } from "../types/client.types";
 import { uploadToCloudinary } from "../api/auth/requests/updateProfileImage";
 import { useToast } from "@/context/ToastContext";
 import { useTranslations } from "next-intl";
