+++ src/lib/types/auth.type.ts
