--- src/lib/types/designated-request.types.ts
@@ -1,13 +1,0 @@
-// src/
-// └── lib/
-//     └── types/
-//         ├── client.types.ts
-//         ├── mover.types.ts
-//         ├── request.types.ts
-//         ├── designated-request.types.ts
-//         ├── estimate.types.ts
-//         ├── review.types.ts
-//         ├── favorite.types.ts
-//         ├── notification.types.ts
-//         ├── region.types.ts
-//         └── index.ts         # 통합 export
