+++ src/lib/types/estimateCalculator.type.ts
