+++ src/lib/types/favorite.type.ts
