+++ src/lib/types/firebase.type.ts
