+++ src/lib/types/index.ts
@@ -1,11 +1,11 @@
-export * from "./auth.types";
-export * from "./mover.types";
-export * from "./review.types";
-export * from "./request.types";
-export * from "./estimate.types";
-export * from "./mover.types";
-export * from "./client.types";
-export * from "./notification.types";
-export * from "./favorite.types";
-export * from "./estimate-calculator.types";
-export * from "./firebase.types";
+export * from "./auth.type";
+export * from "./client.type";
+export * from "./community.type";
+export * from "./estimateCalculator.type";
+export * from "./estimate.type";
+export * from "./favorite.type";
+export * from "./firebase.type";
+export * from "./mover.type";
+export * from "./notification.type";
+export * from "./request.type";
+export * from "./review.type";
