+++ src/lib/types/mover.type.ts
@@ -1,5 +1,5 @@
 import { Control, FieldError, Path, UseFormRegister } from "react-hook-form";
-import { User, Mover } from "./auth.types";
+import { Mover, User } from "./auth.type";
 
 //바로 아래+기사님 프로필 컴포넌트에서 사용
 export type FieldValue = string | string[] | File | undefined;
